--- a/binding.gyp
+++ b/binding.gyp
@@ -11,7 +11,6 @@
         "src/node.cc",
         "src/parser.cc",
         "src/query.cc",
-        "src/query_cursor.cc",
         "src/tree.cc",
         "src/tree_cursor.cc",
         "src/util.cc",