--- a/src/binding.cc
+++ b/src/binding.cc
@@ -3,6 +3,8 @@
 #include "./language.h"
 #include "./node.h"
 #include "./parser.h"
+#include "./query.h"
+#include "./query_cursor.h"
 #include "./tree.h"
 #include "./tree_cursor.h"
 #include "./conversions.h"
@@ -16,6 +18,8 @@ void InitAll(Local<Object> exports) {
   node_methods::Init(exports);
   language_methods::Init(exports);
   Parser::Init(exports);
+  Query::Init(exports);
+  QueryCursor::Init(exports);
   Tree::Init(exports);
   TreeCursor::Init(exports);
 }