Hey,

I need access to the query API for a project, here is a draft implementation of it. Are you interesting in including it in the main repo?

Thanks.

Fixes #56