@maxbrunsfeld 2020-05-04
Hi, sorry for the delay in response. Nice work on this so far.

I would definitely be open to merging this. I think that I would prefer for the API to match the the query API in the WASM binding. Unfortunately, it's not really documented right now; the best way to understand it is to read the [test file](https://github.com/tree-sitter/tree-sitter/blob/master/lib/binding_web/test/query-test.js).

Summary:

1. `Language.query(source)` returns a `Query`.
2. `query.matches` returns an array of matches (using `ts_query_cursor_next_match`)
3. `query.captures` returns an array of captures (using `ts_query_cursor_next_capture`)
4. The `TSQueryCursor` is not exposed directly. There is just a private global cursor that is used to implement `matches` and `captures`.

Thoughts?