@maxbrunsfeld 2020-05-05
Also, I'm just curious if you tried using the wasm-based bindings (`web-tree-sitter`) instead of this module for your project, and if so, what features of this module you found more useful.