@romgrk 2020-05-06
> Also, I'm just curious if you tried using the wasm-based bindings (`web-tree-sitter`) instead of this module for your project, and if so, what features of this module you found more useful.

I haven't tried the wasm version. I remember seeing somewhere that the performance of the wasm version was slower than the native one, thus why I'm interested in this one. If that claim is not applicable anymore I could try the wasm bindings.

I'll reply to the other points, I should have more time in a few days.