@thibaultdalban 2020-05-20
Hello,

I'm also very interested in the query API, and I will be happy to help!
I also use this module and not `web-tree-sitter` for the performance reason mentioned [here](https://github.com/tree-sitter/tree-sitter/blob/master/lib/binding_web/README.md#running-wasm-in-nodejs)