@thibaultdalban 2020-05-22
Unfortunately, I was unable to load the Ruby Language using wasm.
I opened an issue [here](https://github.com/tree-sitter/tree-sitter-ruby/issues/116) about that.