@romgrk 2020-05-26
Hey @maxbrunsfeld, sorry for the delay, there is never enough time to do everything :/

So about your comments, here are my thoughts:

> The `TSQueryCursor` is not exposed directly. There is just a private global cursor that is used to implement `Query.matches` and `Query.captures`.

Sounds good to me, nothing is gained by exposing the query cursor to the users.

> `Query.matches` returns an array of matches (using `ts_query_cursor_next_match`)
> `Query.captures` returns an array of captures (using `ts_query_cursor_next_capture`)

Sounds good as well, though I would still keep `Query.exec(tree: TSTree, callback: (...) -> void)`. I like the `.matches` and `.captures` versions because it avoids crossing the JS/native boundary too much, but OTOH `.exec` can avoid unnecessary JS object allocations if the matches are used right away.
For example, in my use-case the highlights are passed to a [`GtkTextView`](https://developer.gnome.org/gtk3/stable/GtkTextView.html) so the JS objects are never kept around.
All this to say I'm not sure enough which style of API would be more efficient, and I would like to be able to have both possibilities as an user of `node-tree-sitter`. Is it ok with you if we keep both styles?

> `Language.query(source)` returns a `Query.`

On this one I think I'd prefer to keep `new Query(Language, source)`. If I understand correctly, `Language` is a different package. I'm not sure I see how we could practically modify the `Language` to include `.query` on it. This package exposes the `Query` object, therefore it should be the one on which the `Query` creation function should be exposed.


Let me know what you think on the above comments and I'll try to finish the PR in the next week.