@maxbrunsfeld 2020-05-26
> On this one I think I'd prefer to keep `new Query(Language, source)`. If I understand correctly, `Language` is a different package.

Yes, you are absolutely right!

And 👍 to all of your suggestions. 

> Predicates are not handled by tree-sitter right? Do they need to be exposed to the user?

That's right. Ideally, we would handle the `eq?` and `match?` predicates automatically, filtering out matches/captures as needed, but I think we can add that later.