@romgrk 2020-05-29
So the way the node marshalling is setup, when sending a list of nodes from C++ to JS, C++ for each nodes returns either the nodeTypeID or the cached node object. However the queries API does something that doesn't happen otherwise: it can return the same node multiple times (if multiple pattern match). So C++ sends multiple times the same node (serialized in `nodeTransferArray`), JS does the NodeClass instantiation, then tries to store this new instance in C++'s cache via `tree._cacheNode(node)`, but then Tree complains because its assertion isn't respected:

```c
assert(!tree->cached_nodes_.count(key)); // tree.cc : 230  -_-
```

There were a couple ways to solve it, I'm quite hesitant about my implementation, I'd like your input. In JS's `unmarshalNodes`, I've introduced a cache that remembers the nodes that have been converted for this call of `unmarshalNodes`. The cache is passed to `unmarshalNode`, which extracts the node's id as a `BigInt` and uses it as the cache key to avoid recrating the same node multiple times. It's all contained in [this commit](https://github.com/tree-sitter/node-tree-sitter/pull/62/commits/90a7177b4106971e7d84cc5a0019833c0689b4ac).

A few other solutions:
 - Expose a `Local<Object> Tree::GetCachedNode(...)` that would return already cached nodes.
 - Do the Node instantiation in C++
 - Ignore duplicate nodes in `Tree::CacheNode(...)`, let 'em be GC'ed eventually.

Could you let me know what you think of it? I'll gladly rework this.

Other questions/comments:
 - I saw multiple checks for `if (node.id)`; in which case can `node.id` be empty?
 - About predicates, I saw [this](https://github.com/tree-sitter/tree-sitter/blob/master/lib/binding_web/binding.js#L746), I think I'd like to use a similar approach and handle them in JS