@maxbrunsfeld 2020-06-01
> There were a couple ways to solve it, I'm quite hesitant about my implementation, I'd like your input.

I think your solution is pretty good. It looks like `BigInt` was available even in node 10.x, so I think it is fine to use.

> Expose a `Local<Object> Tree::GetCachedNode(...)` that would return already cached nodes.

I am also open to this solution if you think it would be cleaner.

> Do the Node instantiation in C++

I don't want to do this because instantiating objects in C++ is much slower than doing so in JavaScript. It essentially always goes through the slowest-possible path (a hash-table like interface, instead of using hidden classes).

> Ignore duplicate nodes in Tree::CacheNode(...), let 'em be GC'ed eventually.

But how would you prevent the `.matches`/`.captures` methods from multiple different `Node` instances that represent the same underlying node? I think we need to deal with duplicates *explicitly* in order to prevent this.

> I saw multiple checks for `if (node.id)`; in which case can node.id be empty?

This means that Tree-sitter API failed to return a node. For example, if you call `ts_node_child` on a node with no children, it will return a "null" node whose `id` is `NULL`.

> About predicates, I saw this, I think I'd like to use a similar approach and handle them in JS

That sounds great!