@romgrk 2020-06-23
Ok, so I've done the changes, everything should be good now.

I've kept the transfer in a flattened JS array instead of a C buffer. Two minor optimizations, first when `unmarshalNodes` is called I've added `Tree::CacheNodes` to cache multiple nodes at once rather than call `Tree::CacheNode` multiple times in a row.

I've ran some tests (parsing `react-dom.min.js`), this removed about 20ms (on 145ms) from `unmarshalNodes`.

Second optimization is that I don't pass `capture_count` in the buffer anymore, so the transfer array looks like:
```js
  // first match
  5, // pattern index
  "foo",
  "bar,

  // second match
  7, // pattern index
  "baz,
]
```
And the captures are accumulated with
```js
while (typeof returnedMatches[i] === 'string') // ...
```
Didn't expect it to affect the result that much but this gives a ~9% gain (for `Query.prototype.captures`):
```
with-capture-count:    470.35ms (+- 63.07ms)
without-capture-count: 430.4485 (+- 59.45ms) -9.2%
```